lkl: add rumpuser hypercall backend
to be implemented
- libbmk w/ pcimap (for DMA devices)
- musl libc integration with rumprun
- virtio integration with rexec
- outsource /dev/stdin/out part to rexec
- support O= build

FIXME: should remove/move test code somewhere else in kernel/rump.c

lkl: add dummy_irq_chip
lkl: add read_persistent_clock() to override clock value
lkl: arm build
lkl: frankenlibc build with musl libc

+++ tools/lkl/tests/boot.sh
@@ -10,6 +10,12 @@ if [ -z "$fstype" ]; then
     fstype="ext4"
 fi
 
+MKFS=`which mkfs.$fstype 2> /dev/null` || echo -n ""
+if [ -z ${MKFS} ]; then
+    echo "test skip: mkfs.$fstype"
+    exit 0
+fi
+
 file=`mktemp`
 dd if=/dev/zero of=$file bs=1024 count=20480
